% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cut.R
\name{auto_cut}
\alias{auto_cut}
\title{Discretizes a Vector}
\usage{
auto_cut(x, breaks = NULL, n_bins = 27, return_distinct = FALSE,
  x_name = "value", level_name = "level", ...)
}
\arguments{
\item{x}{A vector.}

\item{breaks}{An optional vector of breaks. Only relevant for numeric \code{x}.}

\item{n_bins}{If \code{x} is numeric and no breaks are provided, this is the maximum number of bins allowed or to be created.}

\item{return_distinct}{If TRUE, the resulting \code{data.frame} is deduplicated.}

\item{x_name}{Column name with the values of \code{x} in the output.}

\item{level_name}{Column name with the bin labels of \code{x} in the output.}

\item{...}{Further arguments passed to \code{cut3}.}
}
\value{
A \code{data.frame} with colums \code{x_name} and \code{level_name} and an attribute "breaks" with cut breaks or NULL if none were used.
}
\description{
This function takes a vector \code{x} and returns a \code{data.frame} with two columns. In case \code{x} is not numeric, both columns are identical to the input \code{x}. The same holds if \code{x} is numeric, no \code{breaks} are provided and the number of disjoint values is not larger than \code{n_bins}. Otherwise, \code{x} is discretized by \code{breaks} or by \code{n_bins} quantile cuts. In these cases, the first column of the resulting \code{data.frame} contains the break average (high+low)/2 and the second column the level names of the form "(low, high)". The construction of the level names can be controlled by passing ... arguments to \code{formatC}.
}
\examples{
out <- auto_cut(1:10, n_bins = 3)
auto_cut(LETTERS[1:4], n_bins = 2)
auto_cut(990:1100, n_bins = 3, big.mark = "'", format = "fg")
auto_cut(990:1100, n_bins = 3, big.mark = "'", format = "fg", return_distinct = TRUE)
auto_cut(c(0.0001, 0.0002, 0.0003, 0.005), n_bins = 3, format = "fg")
}
