% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_breaks.R
\name{common_breaks}
\alias{common_breaks}
\title{Common Breaks for multiflashlight}
\usage{
common_breaks(x, v, data, breaks, n_bins, cut_type)
}
\arguments{
\item{x}{An object of class \code{flashlight}.}

\item{v}{The variable to be profiled.}

\item{data}{A \code{data.frame}.}

\item{breaks}{Cut breaks for a numeric \code{v}.}

\item{n_bins}{Maxmium number of unique values to evaluate for numeric \code{v}.}

\item{cut_type}{Cut type}
}
\value{
A vector of breaks
}
\description{
Internal function used to find common breaks from different flashlights.
}
