% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_flashlight.R
\name{is.flashlight}
\alias{is.flashlight}
\alias{is.multiflashlight}
\alias{is.light}
\alias{is.light_performance}
\alias{is.light_performance_multi}
\alias{is.light_importance}
\alias{is.light_importance_multi}
\alias{is.light_breakdown}
\alias{is.light_breakdown_multi}
\alias{is.light_ice}
\alias{is.light_ice_multi}
\alias{is.light_profile}
\alias{is.light_profile_multi}
\alias{is.light_effects}
\alias{is.light_effects_multi}
\alias{is.shap}
\alias{is.light_scatter}
\alias{is.light_scatter_multi}
\alias{is.light_global_surrogate}
\alias{is.light_global_surrogate_multi}
\title{Check functions for flashlight Classes}
\usage{
is.flashlight(x)

is.multiflashlight(x)

is.light(x)

is.light_performance(x)

is.light_performance_multi(x)

is.light_importance(x)

is.light_importance_multi(x)

is.light_breakdown(x)

is.light_breakdown_multi(x)

is.light_ice(x)

is.light_ice_multi(x)

is.light_profile(x)

is.light_profile_multi(x)

is.light_effects(x)

is.light_effects_multi(x)

is.shap(x)

is.light_scatter(x)

is.light_scatter_multi(x)

is.light_global_surrogate(x)

is.light_global_surrogate_multi(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A logical vector of length one.
}
\description{
Checks if an object inherits specific class relevant for the flashlight package.
}
\section{Functions}{
\itemize{
\item \code{is.multiflashlight}: Check for multiflashlight object.

\item \code{is.light}: Check for light object.

\item \code{is.light_performance}: Check for light_performance object.

\item \code{is.light_performance_multi}: Check for light_performance_multi object.

\item \code{is.light_importance}: Check for light_importance object.

\item \code{is.light_importance_multi}: Check for light_importance_multi object.

\item \code{is.light_breakdown}: Check for light_breakdown object.

\item \code{is.light_breakdown_multi}: Check for light_breakdown_multi object.

\item \code{is.light_ice}: Check for light_ice object.

\item \code{is.light_ice_multi}: Check for light_ice_multi object.

\item \code{is.light_profile}: Check for light_profile object.

\item \code{is.light_profile_multi}: Check for light_profile_multi object.

\item \code{is.light_effects}: Check for light_effects object.

\item \code{is.light_effects_multi}: Check for light_effects_multi object.

\item \code{is.shap}: Check for shap object.

\item \code{is.light_scatter}: Check for light_scatter object.

\item \code{is.light_scatter_multi}: Check for light_scatter_multi object.

\item \code{is.light_global_surrogate}: Check for light_global_surrogate object.

\item \code{is.light_global_surrogate_multi}: Check for light_global_surrogate_multi object.
}}

\examples{
a <- flashlight(label = "a")
is.flashlight(a)
is.flashlight("a")
}
