% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_breakdown.R
\name{light_breakdown}
\alias{light_breakdown}
\alias{light_breakdown.default}
\alias{light_breakdown.flashlight}
\alias{light_breakdown.multiflashlight}
\title{Variable Contribution Breakdown for Single Observation}
\usage{
light_breakdown(x, ...)

\method{light_breakdown}{default}(x, ...)

\method{light_breakdown}{flashlight}(x, new_obs, data = x$data,
  by = x$by, v = NULL, order_by_importance = TRUE, top_m = Inf,
  n_max = Inf, seed = NULL, use_linkinv = FALSE,
  after_name = "after", before_name = "before", label_name = "label",
  variable_name = "variable", step_name = "step",
  description_name = "description", digits = 3, ...)

\method{light_breakdown}{multiflashlight}(x, ...)
}
\arguments{
\item{x}{An object of class \code{flashlight} or \code{multiflashlight}.}

\item{...}{Further arguments passed to \code{prettyNum} to format numbers in description text.}

\item{new_obs}{One single new observation to calculate variable attribution for. Needs to be a \code{data.frame} of same structure as \code{data}.}

\item{data}{An optional \code{data.frame}.}

\item{by}{An optional vector of column names used to filter \code{data} for rows with equal values in "by" variables as \code{new_obs}.}

\item{v}{Vector of variables to assess contribution for. Defaults to all except those specified by "y", "w" and "by".}

\item{order_by_importance}{Logical flag indicated if \code{v} should be ordered automatically based on individual contribution (see description).}

\item{top_m}{Maximum number of steps/variable contributions to be calculated.}

\item{n_max}{Maximum number of rows in \code{data} to consider. Set to lower value if \code{data} is large.}

\item{seed}{An integer random seed used to shuffle rows if \code{n_max} is smaller than the number of rows in \code{data}.}

\item{use_linkinv}{Should retransformation function be applied? Default is \code{FALSE}.}

\item{after_name}{Column name in resulting \code{data} containing prediction after the step in \code{step_name}. Defaults to "after".}

\item{before_name}{Column name in resulting \code{data} containing prediction before the step in \code{step_name}. Defaults to "before".}

\item{label_name}{Column name in resulting \code{data} containing the label of the flashlight. Defaults to "label".}

\item{variable_name}{Column name in resulting \code{data} containing the variable names. Defaults to "variable".}

\item{step_name}{Column name in resulting \code{data} containing the step of the prediction process. Defaults to "step".}

\item{description_name}{Column name in resulting \code{data} containing the description text to be visualized. Defaults to "description".}

\item{digits}{Passed to \code{prettyNum} to format numbers in description text.}
}
\value{
An object of class \code{light_breakdown}, \code{light} (and a list) with the following elements.
\itemize{
  \item \code{data} A tibble with results. Can be used to build fully customized visualizations.
  \item \code{by} Same as input \code{by}.
  \item \code{after_name} Same as input \code{after_name}.
  \item \code{before_name} Same as input \code{before_name}.
  \item \code{label_name} Same as input \code{label_name}.
  \item \code{variable_name} Same as input \code{variable_name}.
  \item \code{step_name} Same as input \code{step_name}.
  \item \code{description_name} Same as input \code{description_name}.
}
}
\description{
Calculates sequential variable contribution to the prediction of a single observation, see Gosiewska and Biecek [1]. For non-additive models, the order of the variables visited is relevant and is determined either by the specified order in \code{v} or the non-sequential contribution of each \code{v}. In the first (non-default) case, complexity in the length p of \code{v} is \code{min(n_max, p)}, while in the second case its \code{p + min(n_max, p)}.
}
\details{
The breakdown algorithm assigns the following attribution to the first variable (x): In the query \code{data}, all values in x are set to the value of x in the new observation. The change in the (weighted) average predicted value on \code{data} is the contribution of x on the prediction. This procedure is iterated over \code{v}. To determine the order of the sequence, the same approach is calculated first without iteration, i.e. starting with the original \code{data} for each variable. The approach is described in detail in [1]. If the interative algorithm would be repeated for all possible permutations of the input variables, the resulting attributions would equal SHAP values. Note that the minimum required elements in the (multi-) flashlight are "y", "predict_function", "model", and "data". The latter can also directly be passed to \code{light_breakdown}. Note that by default, no retransformation function is applied.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method not implemented yet.

\item \code{flashlight}: Variable attribution to single observation for a flashlight.

\item \code{multiflashlight}: Variable attribution to single observation for a multiflashlight.
}}

\examples{
fit_part <- lm(Sepal.Length ~ Petal.Length, data = iris)
fit_full <- lm(Sepal.Length ~ ., data = iris)
mod_full <- flashlight(model = fit_full, label = "full", data = iris, y = "Sepal.Length")
mod_part <- flashlight(model = fit_part, label = "part", data = iris, y = "Sepal.Length")
mods <- multiflashlight(list(mod_full, mod_part), by = "Species")
light_breakdown(mod_full, new_obs = iris[1, ])
light_breakdown(mods, new_obs = iris[1, ])
light_breakdown(mods, new_obs = iris[1, ], top_m = 2)

ir <- iris
ir$log_sl <- log(ir$Sepal.Length)
fit_lm <- lm(log_sl ~ Petal.Length, data = ir)
fit_glm <- glm(Sepal.Length ~ Petal.Length, data = ir, family = Gamma(link = log))
fl_lm <- flashlight(model = fit_lm, label = "lm", y = "log_sl", linkinv = exp)
fl_glm <- flashlight(model = fit_glm, label = "glm", y = "Sepal.Length",
  predict_function = function(m, X) predict(m, X, type = "response"))
fls <- multiflashlight(list(fl_lm, fl_glm), data = ir)
light_breakdown(fls, new_obs = ir[1, ])$data
light_breakdown(fls, new_obs = ir[1, ], use_linkinv = TRUE)$data
}
\references{
[1] A. Gosiewska and P. Biecek (2019). IBREAKDOWN: Uncertainty of model explanations for non-additive predictive models. ArXiv <arxiv.org/abs/1903.11420>.
}
\seealso{
\code{\link{plot.light_breakdown}}.
}
