% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_deprecated.R, R/light_recode.R
\name{light_recode}
\alias{light_recode}
\title{DEPRECATED}
\usage{
light_recode(...)

light_recode(...)
}
\arguments{
\item{...}{Deprecated.}
}
\value{
Error message.

Deprecated.
}
\description{
DEPRECATED

Recode Factor Columns - DEPRECATED
}
