% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoints.R
\name{midpoints}
\alias{midpoints}
\title{Midpoints}
\usage{
midpoints(breaks)
}
\arguments{
\item{breaks}{Numeric vector of cut points or a single number specifying the number of intervals desired.}
}
\value{
Vector of the same length as x minus 1 with midpoints of breaks.
}
\description{
Takes a vector of breaks and calculates midpoints of subsequent unique breaks.
}
\examples{
midpoints(1:4)
midpoints(c(4, 4:1))
}
