% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_light_effects.R
\name{plot.light_effects}
\alias{plot.light_effects}
\title{Visualize Different Types of Profiles Together}
\usage{
\method{plot}{light_effects}(x, use = c("response", "predicted", "pd"),
  zero_counts = TRUE, size_factor = 1, facet_scales = "free_x",
  rotate_x = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{light_effects}.}

\item{use}{A vector with elements in "response", "predicted" and "pd" (partial dependence) with elements to show. Defaults to all.}

\item{zero_counts}{Logical flag if 0 count levels should be shown on the x axis.}

\item{size_factor}{Factor used to enlarge default \code{size} in \code{geom_point} and \code{geom_line}.}

\item{facet_scales}{Scales argument passed to \code{facet_wrap}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{...}{Further arguments passed to geoms.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
Visualizes response-, prediction-, and partial dependence profiles of a (multi-)flashlight with respect to a covariable \code{v}. Different flashlights or a single flashlight with one "by" variable are separated by a facet wrap.
}
\examples{
\dontrun{
fit_full <- lm(Sepal.Length ~ ., data = iris)
fit_part <- glm(Sepal.Length ~ Petal.Length, data = iris)
mod_full <- flashlight(model = fit_full, label = "full", data = iris,
  y = "Sepal.Length", w = "Petal.Length")
mod_part <- flashlight(model = fit_part, label = "part", data = iris,
  y = "Sepal.Length", w = "Petal.Length")
mods <- multiflashlight(list(mod_full, mod_part))

plot(light_effects(mod_full, v = "Species"))
x <- light_effects(mod_full, v = "Petal.Width")
plot(x)
plot(x, use = "response")
plot(x, use = "predicted")
plot(x, use = "pd")
plot_counts(plot(x), x, alpha = 0.2)
plot_counts(plot(x, use = "response"), x, alpha = 0.2)
plot_counts(plot(x, use = "pd"), x, alpha = 0.2)

x <- light_effects(mod_full, v = "Petal.Width", stats = "quartiles")
plot(x)
plot(x, size_factor = 3)
plot_counts(plot(x), x, alpha = 0.2)
plot_counts(plot(x, use = "response"), x, alpha = 0.2)

x <- light_effects(mod_full, v = "Petal.Width", by = "Species")
plot(x) +
  scale_color_viridis_d(begin = 0.2, end = 0.8)
p <- plot(x, zero_counts = FALSE)
plot_counts(p, x, zero_counts = FALSE, alpha = 0.2)

x <- light_effects(mod_full, v = "Petal.Width", by = "Species", stats = "quartiles")
plot(x)
plot_counts(plot(x), x, alpha = 0.2)

# Smuggle additional line into the pic
x_add <- light_profile(mod_full, v = "Petal.Width", type = "response", by = "Species")
x_add$data$type <- "observed"
x$predicted <- bind_rows(x$predicted, x_add$data)
(p <- plot(x))
plot_counts(p, x, alpha = 0.2)

x <- light_effects(mods, v = "Petal.Width")
plot(x, zero_counts = TRUE)
plot_counts(plot(x, zero_counts = TRUE), x, alpha = 0.2, zero_counts = TRUE)
plot(light_effects(mods, v = "Petal.Width", stats = "quartiles"))

# Different data: Use eigher fixed breaks or v_labels = FALSE
mod_full <- flashlight(model = fit_full, label = "full",
  data = iris[1:75, ], y = "Sepal.Length")
mod_part <- flashlight(model = fit_part, label = "part",
  data = iris[76:150, ], y = "Sepal.Length")
mods <- multiflashlight(list(mod_full, mod_part))

plot(light_effects(mods, v = "Petal.Length", breaks = 0:8))
eff <- light_effects(mods, v = "Petal.Length")
plot_counts(plot(eff), eff, show_labels = FALSE)
plot_counts(plot(eff, zero_counts = FALSE), eff, zero_counts = FALSE)

# Log-linear OLS vs. Gamma
ir <- iris
ir$log_sl <- log(ir$Sepal.Length)
fit_lm <- lm(log_sl ~ Petal.Length + Petal.Width, data = ir)
fit_glm <- glm(Sepal.Length ~ Petal.Length + Petal.Width,
  data = ir, family = Gamma(link = log))
fl_lm <- flashlight(model = fit_lm, label = "lm", y = "log_sl", linkinv = exp)
fl_glm <- flashlight(model = fit_glm, label = "glm", y = "Sepal.Length",
  predict_function = function(m, X) predict(m, X, type = "response"))
fls <- multiflashlight(list(fl_lm, fl_glm), data = ir)

plot(light_effects(fls, v = "Petal.Length"))
plot(light_effects(fls, v = "Petal.Length", use_linkinv = FALSE))
}
}
\seealso{
\code{\link{light_effects}}, \code{\link{plot_counts}}.
}
