% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_ice.R
\name{plot.light_ice}
\alias{plot.light_ice}
\title{Visualize ICE profiles}
\usage{
\method{plot}{light_ice}(x, facet_scales = "fixed", rotate_x = FALSE, ...)
}
\arguments{
\item{x}{An object of class "light_ice".}

\item{facet_scales}{Scales argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{rotate_x}{Should x axis labels be rotated by 45 degrees?}

\item{...}{Further arguments passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}
}
\value{
An object of class "ggplot".
}
\description{
Minimal visualization of an object of class "light_ice" as \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.
The object returned is of class "ggplot" and can be further customized.
}
\details{
Each observation is visualized by a line. The first "by" variable is represented
by the color, a second "by" variable or a multiflashlight by facets.
}
\seealso{
\code{\link[=light_ice]{light_ice()}}
}
