% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_flashlight.R
\name{predict.flashlight}
\alias{predict.flashlight}
\title{Predictions for flashlight}
\usage{
\method{predict}{flashlight}(object, ...)
}
\arguments{
\item{object}{An object of class \code{flashlight}.}

\item{...}{Arguments used to update the flashlight.}
}
\value{
A vector with predictions.
}
\description{
Predict method for an object of class \code{flashlight}. Pass additional elements to update the flashlight, typically \code{data}.
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
(fl <- flashlight(model = fit, data = iris, y = "Sepal.Length", label = "ols"))
predict(fl)[1:5]
predict(fl, data = iris[1:5, ])
predict(fl, data = iris[1:5, ], linkinv = exp)
}
