% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.light}
\alias{print.light}
\title{Prints light Object}
\usage{
\method{print}{light}(x, ...)
}
\arguments{
\item{x}{A on object of class "light".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for an object of class "light".
}
\examples{
fit <- lm(Sepal.Length ~ ., data = iris)
fl <- flashlight(model = fit, label = "lm", y = "Sepal.Length", data = iris)
light_performance(fl, v = "Species")
}
