% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_small.R
\name{zap_small}
\alias{zap_small}
\title{zap_small}
\usage{
zap_small(x, eps = 1e-12, val = 0)
}
\arguments{
\item{x}{A vector.}

\item{eps}{Values closer to 0 than this value are being replaced by \code{val}.}

\item{val}{Value to replace bad values.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
Internal function used to clip small values.
}
